Deck name exists wip round 2

+++ be/app/src/main/java/com/deckbop/api/service/DeckService.java
@@ -7,6 +7,7 @@ import com.deckbop.api.data.SQLTemplates;
 import com.deckbop.api.data.dao.impl.DeckDatabaseDAO;
 import com.deckbop.api.data.dao.impl.UserDatabaseDAO;
 import com.deckbop.api.exception.CreateDeckException;
+import com.deckbop.api.exception.DeckNameExistsException;
 import com.deckbop.api.model.Card;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
@@ -26,23 +27,26 @@ public class DeckService  {
     @Autowired
     LoggingService loggingService;
 
-    public void createDeck(DeckRequest request) throws CreateDeckException {
-        try {
-            String deckName = request.getName();
-            long userId = request.getUserId();
-            Long deck_id = deckDatabaseDAO.insertDeck(deckName, userId);
-            if (deck_id > 0) {
-                String cardValues = this.getCardDataSQL(request, deck_id);
-                deckDatabaseDAO.addCardsToDeck(SQLTemplates.addCardsToDeck + cardValues);
+    public void createDeck(DeckRequest request) throws CreateDeckException, DeckNameExistsException {
+        if (validateDeckRequest(request)) {
+            try {
+                String deckName = request.getName();
+                long userId = request.getUserId();
+                Long deck_id = deckDatabaseDAO.insertDeck(deckName, userId);
+                if (deck_id > 0) {
+                    String cardValues = this.getCardDataSQL(request, deck_id);
+                    deckDatabaseDAO.addCardsToDeck(SQLTemplates.addCardsToDeck + cardValues);
 
-            } else {
-                loggingService.error(this,"SQL ERROR could not create deck.");
-                throw new CreateDeckException("SQL ERROR could not create deck.");
+                } else {
+                    loggingService.error(this, "SQL ERROR could not create deck.");
+                    throw new CreateDeckException("SQL ERROR could not create deck.");
+                }
+            } catch (DataAccessException e) {
+                loggingService.error(this, "SQL error while creating deck");
+                throw e;
             }
-        }
-        catch (DataAccessException e) {
-            loggingService.error(this, "SQL error while creating deck");
-            throw e;
+        } else {
+            throw new DeckNameExistsException("Could not update, deck name exists");
         }
     }
 
@@ -80,16 +84,19 @@ public class DeckService  {
         return Optional.ofNullable(deck);
     }
 
-    public void updateDeck(DeckRequest request, long deck_id) {
-        try {
-            deckDatabaseDAO.updateDeckTable(request.getName(), deck_id);
-            deckDatabaseDAO.deleteCardsFromDeck(deck_id);
-            String cardValues = this.getCardDataSQL(request, deck_id);
-            deckDatabaseDAO.addCardsToDeck(SQLTemplates.addCardsToDeck + cardValues);
-        }
-        catch (Exception e) {
-            loggingService.error(this,"SQL error in updateDeck");
-            throw e;
+    public void updateDeck(DeckRequest request, long deck_id) throws DeckNameExistsException {
+        if (validateDeckRequest(request)) {
+            try {
+                deckDatabaseDAO.updateDeckTable(request.getName(), deck_id);
+                deckDatabaseDAO.deleteCardsFromDeck(deck_id);
+                String cardValues = this.getCardDataSQL(request, deck_id);
+                deckDatabaseDAO.addCardsToDeck(SQLTemplates.addCardsToDeck + cardValues);
+            } catch (Exception e) {
+                loggingService.error(this, "SQL error in updateDeck");
+                throw e;
+            }
+        } else {
+            throw new DeckNameExistsException("Could not update, deck name exists");
         }
     }
 
@@ -140,5 +147,9 @@ public class DeckService  {
             values[i] =  "(" + deckId + ", '" + cardId + "', " + cardQuantity + ")";
         }
         return String.join(", ", values);
+    }
+
+    private boolean validateDeckRequest(DeckRequest request){
+        return !deckDatabaseDAO.checkDeckInDeckTable(request.getUserId(), request.getName()).next();
     }
 }
